Add files via upload
Handwritten digit recognition has recently been of very interest among the researchers because of the evolution of various Artificial learning, Deep Learning and Computer Vision algorithms. It is the ability of a computer system to recognize the handwritten inputs like. In this report, we compare the results of some of the most widely used Algorithms like SVM & KNN with Deep Learning algorithms like multilayer CNN using Keras with Theano and Tensor  flow. Using these, we will be able to get the accuracy of CNN (Keras+Theano) as compared to using SVM and KNN. For our project we are using a MNIST Database, a subset of a larger set NIST which is a database of 70,000 handwritten digits, divided into 60,000 training examples and 10,000 testing samples. The images in the MNIST dataset are present in the form of an array consisting of 28x28 values representing an image along with their labels. This is also the same in case of the testing images. The data is stored in four files
1.train-images-idx3-ubyte: training set images
2. Train-labels-idx1-ubyte: training set labels
3. t10k-images-idx3-ubyte: test set images
4. t10k-labels-idx1-ubyte: test set labels